pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aider-gatekeeper"
version = "0.1.0"
description = "A FastAPI proxy to prevent OOM crashes when running Aider with local LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "transformers>=4.35.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "respx>=0.20.0",
]

[project.scripts]
gatekeeper = "aider_gatekeeper.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/aider_gatekeeper"]
